requirements.txt
Flask
werkzeug
jinja2
gunicorn

psycopg
SQLAlchemy
SQLAlchemy-Utils
Flask-SQLAlchemy
alembic==1.13.1

redis
celery

pytest
pytest-cov
flake8
isort
black

flask-debugtoolbar
Flask-Static-Digest
Flask-Secrets
Flask-DB

openai
pyodbc
python_dotenv
langchain
langchain_community
langchain_core
langchain_openai
